+++ zk_stdlib/examples/identity/enrollment.rs
@@ -73,7 +73,7 @@ impl Relation for CredentialEnrollment {
         instance: Value<Self::Instance>,
         witness: Value<Self::Witness>,
     ) -> Result<(), Error> {
-        let secp256k1_curve = std_lib.secp256k1_curve();
+        let secp256k1_curve = std_lib.secp256k1();
         let b64_chip = std_lib.base64();
 
         // Assign the PK as public input.
@@ -142,8 +142,8 @@ impl CredentialEnrollment {
         message: &[AssignedByte<F>],
         sig: Value<ECDSASig>,
     ) -> Result<(), Error> {
-        let secp256k1_curve = std_lib.secp256k1_curve();
-        let secp256k1_scalar = std_lib.secp256k1_scalar();
+        let secp256k1_curve = std_lib.secp256k1();
+        let secp256k1_scalar = secp256k1_curve.scalar_field_chip();
         let secp256k1_base = secp256k1_curve.base_field_chip();
 
         // Assign the message and hash it.